requirements.txt
streamlit
pillow
matplotlib
reportlab
